Tried the screenAsBytes.
When I print it out it at first it seems like it is done the right way,
because when printing out every element in my screen array it will print
00100000 as the first byte, which should be there and should be 32 in
decimal. However, when I print out every element in the binary array
'screen' java will print decimals and instead of printing 32 it somehow
prints -96, but 32 for the corresponding byte in the 'answer' array.